// src/app/models/station.ts
export class Station {
    number: string;
    name: string;
    address: string;
    address2: string;
    commune: string;
    nmarrond: string;
    bonus: string;
    pole: string;
    lat: string;
    lng: string;
    bike_stands: string;
    status: string;
    available_bike_stands: string;
    available_bikes: string;
    availabilitycode: string;
    availability: string;
    banking: string;
    gid: string;
    last_update: string;
    last_update_fme: string;
}